move calculate method to generic solver
